Break idwrap into its own mod as it's getting lengthy
