+++ packages/cadmium/src/message/idwrap/de.rs
@@ -1,63 +1,11 @@
 use std::fmt;
 
 use serde::de::{self, MapAccess, Visitor};
-use serde::ser::SerializeMap;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use tsify_next::Tsify;
+use serde::{Deserialize, Deserializer, Serialize};
 
-use crate::IDType;
+use crate::message::{Identifiable, MessageHandler};
 
-use super::{Identifiable, MessageHandler, ProjectMessageHandler};
-
-#[derive(Tsify, Debug, Clone)]
-#[tsify(from_wasm_abi)]
-pub struct IDWrap<T: Serialize + for<'de> Deserialize<'de> + wasm_bindgen::convert::RefFromWasmAbi> {
-    pub id: u64,
-    pub inner: T,
-}
-
-impl<T: Serialize + for<'de> Deserialize<'de> + wasm_bindgen::convert::RefFromWasmAbi> IDWrap<T> {
-    pub fn new(id: IDType, h: T) -> Self {
-        Self {
-            id,
-            inner: h,
-        }
-    }
-
-    pub fn id(&self) -> IDType {
-        self.id
-    }
-
-    pub fn inner(&self) -> &T {
-        &self.inner
-    }
-}
-
-// First level message handler
-impl<T, U> ProjectMessageHandler for IDWrap<T>
-where
-    T: MessageHandler<Parent = U> + Serialize + for<'de> Deserialize<'de> + wasm_bindgen::convert::RefFromWasmAbi,
-    U: Identifiable<Parent = crate::project::Project>,
-{
-    fn handle_project_message(&self, project: &mut crate::project::Project) -> anyhow::Result<Option<IDType>> {
-        let prnt = U::from_parent_id(project, self.id)?;
-        self.inner.handle_message(prnt)
-    }
-}
-
-// Second level message handler
-impl<T, C, P> MessageHandler for IDWrap<T>
-where
-    T: MessageHandler<Parent = C> + Serialize + for<'de> Deserialize<'de> + wasm_bindgen::convert::RefFromWasmAbi,
-    C: Identifiable<Parent = P>,
-    P: Identifiable<Parent = crate::project::Project>,
-{
-    type Parent = C::Parent;
-    fn handle_message(&self, parent: Self::Parent) -> anyhow::Result<Option<IDType>> {
-        let prnt = C::from_parent_id(&parent, self.id)?;
-        self.inner.handle_message(prnt)
-    }
-}
+use super::IDWrap;
 
 impl<'de, T, C> Deserialize<'de> for IDWrap<T>
 where
@@ -129,31 +77,5 @@ where
         }
 
         deserializer.deserialize_map(IDWrapVisitor::new())
-    }
-}
-
-impl<T, C> Serialize for IDWrap<T>
-where
-    T: MessageHandler<Parent = C> + Serialize + for<'de> Deserialize<'de> + wasm_bindgen::convert::RefFromWasmAbi,
-    C: Identifiable,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        // Create a serializer map with the appropriate capacity
-        let mut map = serializer.serialize_map(Some(1))?;
-        // Add the id field using the name from the Identifiable trait
-        map.serialize_entry(C::ID_NAME, &self.id)?;
-        // Add the inner object fields
-        serde_json::to_value(&self.inner)
-            .map_err(serde::ser::Error::custom)?
-            .as_object()
-            .ok_or_else(|| serde::ser::Error::custom("Expected object"))?
-            .iter().try_for_each(|(k, v)| {
-                map.serialize_entry(k, v)?;
-                Ok(())
-            })?;
-        map.end()
     }
 }
